requirements.txt
fastapi==0.86.0
SQLAlchemy==1.4.41
pydantic==1.10.2
alembic==1.8.1
pydantic[email]
passlib[bcrypt]
types-passlib
uvicorn[standard]
pytest==7.2.0
psycopg2-binary==2.9.5
databases[asyncpg]
databases[aiosqlite]
python-multipart==0.0.5
pytest-asyncio==0.20.2
requests==2.28.1

black
mypy
pylint